changes upto this point

+++ app/controllers/ContractsController.php
@@ -70,7 +70,10 @@ class ContractsController extends \BaseController {
 
 		$employer_notification = $this->notification->notify_employer_about_contract($contract->employer_id, $job->job_title );
 
-		$job->delete();
+		if(check_if_slot_is_not_available($job->id))
+		{
+			$job->delete();
+		}
 
 		return Response::json($contract);
 	}
@@ -89,4 +92,26 @@ class ContractsController extends \BaseController {
 		->withEmployers($employers);
 	}
 
+	function adminpoplulaterenewcontract()
+	{
+		$id = Input::get('contract_id');
+
+		$contract = Contract::withTrashed()->findOrFail($id);
+
+		return Response::json($contract);
+	}
+
+	function adminsaveemployercontractrenew()
+	{
+		$contract = $this->contract->adminsaveemployercontractrenew(Input::all());
+
+		return $contract;
+	}
+
+	function admindeleteemployeecontract($id)
+	{
+		$recruitcontract = $this->contract->admindeleteemployeecontract($id);
+
+		return Redirect::back();
+	}
 }
